requirements.txt
faiss-cpu==1.8.0
flask
langchain==0.1.10
openai==1.13.3
python-dotenv==1.0.1
pypdf==4.0.2
-e .
